% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeTests.R
\docType{data}
\name{melonProbes450k}
\alias{melonProbes450k}
\title{Dataframe containing potentially bad probes for 450K microarray data}
\format{
A data frame with 800,000 rows and 9 variables:
\describe{
  \item{Zhou_HG19}{Logical Vector of CpGs that are marked TRUE in Zhou's General Masking column for HG19}
  \item{Zhou_HG38}{Logical Vector of CpGs that are marked TRUE in Zhou's General Masking column for HG38}
  \item{PercLowVariation}{Numeric Vector describing the number of times a CpG is ranked in bottom 5th percentile of variation}
  \item{PercBadBc}{Numeric Vector describing the number of times a CpG has a beadcount  4 in 5 of samples.}
  \item{PercBadPv}{Numeric Vector describing the number of times a CpG has a detection p value  0.05 in 1 of samples.}	
}
}
\source{
\url{http://zwdzwd.github.io/InfiniumAnnotation}
        
   Pidsley, R., Zotenko, E., Peters, T.J. et al. Critical evaluation of the Illumina MethylationEPIC BeadChip microarray for whole-genome DNA methylation profiling. Genome Biol 17, 208 (2016) doi:10.1186\/s13059-016-1066-1
}
\usage{
data(melonProbes450k)
}
\description{
Dataframe containing potentially bad probes for 450K microarray data
}
\examples{
    data(melonProbes450k)

}
\keyword{datasets}
