% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeTests.R
\docType{methods}
\name{testProbes,RGChannelSetExtended-method}
\alias{testProbes,RGChannelSetExtended-method}
\title{Test Illumina CpG Probes for potential issues}
\usage{
\S4method{testProbes}{RGChannelSetExtended}(betas, manifest = c("450k",
  "EPIC"), beadcounts = NULL, detection = NULL, nb = 0.2, np = 0.2,
  nvar = 0.5, ot, nbCount = 3, nbThresh = 0.05, pvCount = 0.05,
  pvThresh = 0.01, nvarThresh = 0.05)
}
\description{
Test Illumina CpG Probes for potential issues
}
