% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeTests.R
\name{testProbes}
\alias{testProbes}
\title{Test Illumina CpG Probes for potential issues}
\usage{
testProbes(
  betas,
  manifest = c("450k", "EPIC"),
  beadcounts = NULL,
  detection = NULL,
  nb = 0.2,
  np = 0.2,
  nvar = 0.5,
  ot,
  nbCount = 3,
  nbThresh = 0.05,
  pvCount = 0.05,
  pvThresh = 0.01,
  nvarThresh = 0.05
)
}
\arguments{
\item{betas}{A matrix contain beta values, or a RGChannelSetExtended, or a methylumiset}

\item{manifest}{Specify which manifest to use}

\item{beadcounts}{Default: NULL, provide with a matrix containing beadcounts}

\item{detection}{Default: NULL, provide with a matrix containing detection P values}

\item{nb}{A value between 0 and 1, represents the \% of datasets a probe has failed in. Default is 0.2 (20\%) however all probes can be tested by specifing a value of 0.}

\item{np}{A value between 0 and 1, represents the \% of datasets a probe has failed in. Default is 0.2 (20\%) however all probes can be tested by specifing a value of 0.}

\item{nvar}{A value between 0 and 1, represents the \% of datasets a probe has failed in. Default is 0.5 (50\%) however all probes can be tested by specifing a value of 0.}

\item{ot}{A character vector describing the probe design for CpGs (must contain 'I' and 'II')}

\item{nbCount}{The beadcount to test probes in your dataset against}

\item{nbThresh}{The proportion of samples}

\item{pvCount}{The detection p value to test probes in your dataset against}

\item{pvThresh}{The proportion of samples to which a probe can fail in}

\item{nvarThresh}{The percentile of variation a probe can be in.}
}
\value{
A data.frame of nrow(betas) x 3 columns specifying which probes have failed which tests.
}
\description{
Tests Illumina CpGs for potential issues that affect a given proportion of datasets. Useful for removing hits that may be unlikely to be reproduced
}
\examples{
#

}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
